\name{soft}
\alias{soft}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Internal glmlep functions
}
\description{
 Internal glmlep functions
}
\usage{
soft(z, lambda)
}
\arguments{
  \item{z}{
 The partial least square estimate.
}
  \item{lambda}{
 The tuning parameter.
}
}
\details{
These are not intended for use by users.
}
\author{
 Canhong Wen, Hao Lin, Shaoli Wang and Xueqin Wang.

Maintainer: Canhong Wen <wencanhong@gmail.com>
}